Moved state components to page component dirs
