--- src/components/state/errorState/errorState.styles.ts
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const styles = {
-  container: {
-    display: 'flex',
-    justifyContent: 'center',
-    marginTop: '150px',
-  },
-} as { [className: string]: React.CSSProperties };
