+++ src/components/state/errorState/errorState.tsx
@@ -1,7 +1,8 @@
 import {
-  EmptyState as PfEmptyState,
+  EmptyState,
   EmptyStateBody,
   EmptyStateIcon,
+  EmptyStateVariant,
   Title,
 } from '@patternfly/react-core';
 import { ErrorCircleOIcon } from '@patternfly/react-icons/dist/js/icons/error-circle-o-icon';
@@ -9,7 +10,6 @@ import { LockIcon } from '@patternfly/react-icons/dist/js/icons/lock-icon';
 import { AxiosError } from 'axios';
 import React from 'react';
 import { InjectedTranslateProps, translate } from 'react-i18next';
-import { styles } from './errorState.styles';
 
 interface ErrorStateProps extends InjectedTranslateProps {
   error: AxiosError;
@@ -35,15 +35,13 @@ const ErrorStateBase: React.SFC<ErrorStateProps> = ({
   }
 
   return (
-    <div style={styles.container}>
-      <PfEmptyState>
-        <EmptyStateIcon icon={icon} />
-        <Title headingLevel="h2" size="lg">
-          {title}
-        </Title>
-        <EmptyStateBody>{subTitle}</EmptyStateBody>
-      </PfEmptyState>
-    </div>
+    <EmptyState variant={EmptyStateVariant.large} className="pf-m-redhat-font">
+      <EmptyStateIcon icon={icon} />
+      <Title headingLevel="h2" size="lg">
+        {title}
+      </Title>
+      <EmptyStateBody>{subTitle}</EmptyStateBody>
+    </EmptyState>
   );
 };
