+++ src/pages/state/maintenance/maintenance.tsx
@@ -1,7 +1,7 @@
 import { Main } from '@redhat-cloud-services/frontend-components/components/Main';
-import { MaintenanceState } from 'components/state/maintenanceState/maintenanceState';
 import React from 'react';
 import { withRouter } from 'react-router-dom';
+import { MaintenanceState } from './maintenanceState';
 
 const Maintenance = () => {
   return (
