+++ src/pages/state/noProviders/notProviders.tsx
@@ -1,7 +1,7 @@
 import { Main } from '@redhat-cloud-services/frontend-components/components/Main';
-import { NoProvidersState } from 'components/state/noProvidersState/noProvidersState';
 import React from 'react';
 import { withRouter } from 'react-router-dom';
+import { NoProvidersState } from './noProvidersState';
 
 const NoProviders = () => {
   return (
