+++ src/pages/state/notAuthorized/notAuthorized.tsx
@@ -1,7 +1,7 @@
 import { Main } from '@redhat-cloud-services/frontend-components/components/Main';
-import { NotAuthorizedState } from 'components/state/notAuthorizedState/notAuthorizedState';
 import React from 'react';
 import { withRouter } from 'react-router';
+import { NotAuthorizedState } from './notAuthorizedState';
 
 const NotAuthorized = () => {
   return (
